cam has valid getPosition/getX/getY/getDrawCenter functions that work

+++ nova2d/nova-zero/src/graphics/Text.cpp
@@ -106,8 +106,8 @@ namespace novazero
 
 			if (IsFixed())
 			{
-				m_DrawRect.x = GetX() + OffsetX();
-				m_DrawRect.y = GetY() + OffsetY();
+				m_DrawRect.x = (int)(GetX() + OffsetX());
+				m_DrawRect.y = (int)(GetY() + OffsetY());
 			}
 			else
 			{
